// Card_Game_Engine/ClientApp/src/app/_reusable-components/card/namespace/enums/card-name.enum.ts
export enum CardNameEnum {
  ACE = 'Ace',
  TWO = 'Two',
  THREE = 'Three',
  FOUR = 'Four',
  FIVE = 'Five',
  SIX = 'Six',
  SEVEN = 'Seven',
  EIGHT = 'Eight',
  NINE = 'Nine',
  TEN = 'Ten',
  JACK = 'Jack',
  QUEEN = 'Queen',
  KING = 'King',
  JOKER = 'Joker'
}

export enum CardValueEnum {
  Ace = '1',
  Two = '2',
  Three = '3',
  Four = '4',
  Five = '5',
  Six = '6',
  Seven = '7',
  Eight = '8',
  Nine = '9',
  Ten = '10',
  Jack = '11',
  Queen = '12',
  King = '13',
  Joker = '14'
}